delete transport layer. we don't need this level of abstraction yet. I introduce ProtocolEngine which is very similar to transport but the goal is different. it gives two main queue and connect connection layer to controller.

+++ src/main/java/artronics/chapar/queue/DataInOutQueue.java
@@ -7,8 +7,8 @@ public class DataInOutQueue implements DataInOutQueueContract
 {
     private static final int MAX_QUEUE_CAPACITY = 20;
 
-    private static final ArrayBlockingQueue<byte[]> dataInQueue = new ArrayBlockingQueue(MAX_QUEUE_CAPACITY);
-    private static final ArrayBlockingQueue<byte[]> dataOutQueue = new ArrayBlockingQueue(MAX_QUEUE_CAPACITY);
+    private static final ArrayBlockingQueue<int[]> dataInQueue = new ArrayBlockingQueue(MAX_QUEUE_CAPACITY);
+    private static final ArrayBlockingQueue<int[]> dataOutQueue = new ArrayBlockingQueue(MAX_QUEUE_CAPACITY);
 
     @Override
     public ArrayBlockingQueue getDataInQueue()
